RoleRAG-PAIMON/requirements.txt
google-generativeai>=0.3.0
networkx>=3.0
python-igraph>=0.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pydantic>=2.0.0
python-dotenv>=1.0.0
